Add spdlogging and argparse
Description:
In order the keep the project easy to understand, debug and modify,
correct logging and cli argument parsing capabilities are needed

Solution:
After a research of the available open source solutions, spdlogging and
argparse were chosen to fill the described functions. Devenv and
CMakeLists were updated accordingly. The basic structure of logging and
argument parsing is implemented in the main source file