Automatise development environment setup
Description:
It is important to make the development flow as easy as possible,
especially for cpp projects. To achieve this, the setup of the
development environment should be self-contained and automatized as
much as possible.

Solution:
Created CONAN_HOME variable, so that conan settings and downloads can be
found in the local project dir, not globally. Created setup script for
conan and git environment for local project that is executed when
entering the project with direnv.