Add newton-raphson method POC
Description:
Calculating the result of non-linear equation systems is crucial for
geometric constraint solving, since most of the graph decompositions
only result on relatively simple systems, but they still need to be
calculated.

Solution:
Added a proof of concept implementation of the Newton-Raphson method
that later can be updated to be generally useable.