Update Newton-Raphson method to Point-Distance triangles and delete own equations
Description:
As a first step, the most basic constraint system should be solvable: it
only contains points and distance constraints and they create a
well-constrained system. To achieve this it is enough to create a
specific Newton-Raphson method, that is a bit better than a POC
implementation and can be expanded later.

Solution:
Implemented a specific Newton-Raphson method for 2x2 matrices and for
triangle decomposed system. It depends on the eigen and autodiff
libraries. Deleted own equation representation, since it will be dead
code. Autodiff does everything, it is tested, it is faster and it is MIT
licensed.