Fix UUID reseeding
Description:
The UUID generation is based on the PRNG Mt19937. This should be seeded
only once per program run. The current code reseeds it at every function
call. This creates unnecessary calculations and decreases the quality of
the generated random numbers.

Solution:
Fixed the described problem by creating a hidden static  generator object
in a new source file. It is only initialized once at the start of the
program with one seed. This object is used at every generate call.