Create graph interface and default implementation
Description:
In the area of geometric constraint solving graphs are frequently used.
Algorithms from graph theory are often needed. This can create a need
for different graph backends, especially when an algorithm is too
complex to implement ourselves and a third party library is needed to
take place. It is important to create interfaces that make it possible
to switch backends without changing the whole codespace.

Solution:
Created template interfaces constrained by concepts for graphs, edges
and nodes. Provided a default implementation for each of them.