Binding simplification and OGDF fix
Description:
Before the first demo of this project, it should be in a relatively
stable state. The current binding is way too complex, and the internal
state of the OGDF library is still unstable.

Solution:
Implemented a new binding approach with claude-code, where the state is
only stored in C++ and a handler class is out on the binding API. Also,
refactored the ogdf wrapper so that ogdf is only used one it is
explicitly needed, avoiding storing its state. This is slower, but
stable. For now this is enough, since later OGDF should be removed from
the project.